requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
cloudpickle==2.0.0
dask==2022.2.0
fsspec==2022.2.0
idna==3.3
locket==0.2.1
numpy==1.21.5
packaging==21.3
pandas==1.3.5
partd==1.2.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
toolz==0.11.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
